+++ src/tracer/src/cli/commands.rs
