+++ src/tracer/src/cli/handlers/info.rs
@@ -1,3 +1,20 @@
+use crate::daemon::client::DaemonClient;
+use anyhow::Result;
+pub async fn info(api_client: &DaemonClient, json: bool) -> Result<()> {
+    let info = match api_client.send_info_request().await {
+        Ok(info) => info,
+        Err(e) => {
+            let mut display = InfoDisplay::new(80, json);
+            tracing::error!("Error getting info response: {e}");
+            display.print_error();
+            return Ok(());
+        }
+    };
+    let display = InfoDisplay::new(150, json);
+    display.print(info);
+    Ok(())
+}
+
 use crate::daemon::structs::InfoResponse;
 use crate::process_identification::constants::{LOG_FILE, STDERR_FILE, STDOUT_FILE};
 use crate::utils::cli::BoxFormatter;
