+++ src/tracer/src/cli/handlers/mod.rs
@@ -1,0 +1,8 @@
+mod update;
+pub use update::update;
+
+mod info;
+
+pub use info::info;
+mod setup;
+pub use setup::setup;
