+++ src/tracer/src/cli/mod.rs
@@ -1,0 +1,7 @@
+pub mod commands;
+mod handlers;
+mod helper;
+mod process_command;
+pub use process_command::process_command;
+mod process_daemon_command;
+pub mod setup;
