+++ src/tracer/src/cli/setup.rs
