+++ src/tracer/src/lib.rs
@@ -3,18 +3,16 @@ use crate::process_identification::types::cli::{
     params::{FinalizedInitArgs, TracerCliInitArgs},
 };
 
+pub mod cli;
 pub mod client;
 pub mod cloud_providers;
-pub mod commands;
 pub mod config;
 pub mod constants;
 pub mod daemon;
 pub mod extracts;
-pub mod logging;
-pub mod nondaemon_commands;
-pub mod process_command;
 pub mod process_identification;
 pub mod utils;
+
 /// Runs tracer init in interactive mode
 pub fn init_command_interactive_mode(cli_args: TracerCliInitArgs) -> FinalizedInitArgs {
     InteractiveInitArgs::from_partial(cli_args)
