--- src/tracer/src/nondaemon_commands.rs
@@ -1,168 +1,0 @@
-use std::process::Command;
-
-use crate::config::Config;
-use crate::daemon::client::DaemonClient;
-use crate::process_identification::constants::{
-    FILE_CACHE_DIR, PID_FILE, STDERR_FILE, STDOUT_FILE,
-};
-use crate::utils::system_info::{is_root, is_sudo_installed};
-use crate::utils::InfoDisplay;
-use anyhow::{bail, Context, Result};
-use colored::Colorize;
-use std::result::Result::Ok;
-use tokio::time::sleep;
-use tracing::debug;
-
-pub fn clean_up_after_daemon() -> Result<()> {
-    std::fs::remove_file(PID_FILE).context("Failed to remove pid file")?;
-    std::fs::remove_file(STDOUT_FILE).context("Failed to remove stdout file")?;
-    std::fs::remove_file(STDERR_FILE).context("Failed to remove stderr file")?;
-    std::fs::remove_dir_all(FILE_CACHE_DIR).context("Failed to remove cache directory")?;
-    Ok(())
-}
-
-pub async fn wait(api_client: &DaemonClient) -> Result<()> {
-    for n in 0..5 {
-        match api_client.send_info().await {
-            // if timeout, retry
-            Err(e) => {
-                if !(e.is_timeout() || e.is_connect()) {
-                    bail!(e)
-                }
-            }
-            Ok(resp) => {
-                if resp.status().is_success() {
-                    return Ok(());
-                }
-
-                debug!("Got response, retrying: {:?}", resp);
-            }
-        }
-
-        let duration = 1 << n;
-        let attempts = match duration {
-            1 => 1,
-            2 => 2,
-            4 => 3,
-            8 => 4,
-            _ => 5,
-        };
-
-        println!(
-            "Starting daemon... [{:.<20}] ({} second{} elapsed)",
-            ".".repeat(attempts.min(20)),
-            duration,
-            if duration > 1 { "s" } else { "" }
-        );
-        sleep(std::time::Duration::from_secs(duration)).await;
-    }
-
-    bail!("Daemon not started yet")
-}
-pub async fn print_info(api_client: &DaemonClient, json: bool) -> Result<()> {
-    let info = match api_client.send_info_request().await {
-        Ok(info) => info,
-        Err(e) => {
-            let mut display = InfoDisplay::new(80, json);
-            tracing::error!("Error getting info response: {e}");
-            display.print_error();
-            return Ok(());
-        }
-    };
-    let display = InfoDisplay::new(150, json);
-    display.print(info);
-    Ok(())
-}
-
-pub async fn setup_config(
-    api_key: &Option<String>,
-    process_polling_interval_ms: &Option<u64>,
-    batch_submission_interval_ms: &Option<u64>,
-) -> Result<()> {
-    let mut current_config = Config::default();
-    if let Some(api_key) = api_key {
-        current_config.api_key.clone_from(api_key);
-    }
-    if let Some(process_polling_interval_ms) = process_polling_interval_ms {
-        current_config.process_polling_interval_ms = *process_polling_interval_ms;
-    }
-    if let Some(batch_submission_interval_ms) = batch_submission_interval_ms {
-        current_config.batch_submission_interval_ms = *batch_submission_interval_ms;
-    }
-
-    //ConfigLoader::save_config(&current_config)?;
-
-    Ok(())
-}
-
-pub async fn update_tracer() -> Result<()> {
-    // TODO commenting out for now, as we get the s3 main release
-    // let octocrab = octocrab::instance();
-    // let release = octocrab
-    //     .repos(REPO_OWNER, REPO_NAME)
-    //     .releases()
-    //     .get_latest()
-    //     .await?;
-
-    // let current = Version::current_str();
-    // let latest = &release.tag_name;
-
-    // let current_ver: Version = current.parse().ok().unwrap();
-    // let latest_ver: Version = latest.parse().ok().unwrap();
-    //
-    // if latest_ver <= current_ver {
-    //     println!(
-    //         "\nTracer is already at the latest version: {}.",
-    //         current_ver
-    //     );
-    //     return Ok(());
-    // }
-
-    // println!("\nA new version of Tracer is available!");
-    // println!("\nVersion Information:");
-    // println!("  Current Version: {}", current_ver);
-    // println!("  Latest Version:  {}", latest_ver);
-    //
-    // println!("\nWould you like to proceed with the update? [y/N]");
-    // let mut input = String::new();
-    // std::io::stdin().read_line(&mut input)?;
-    //
-    // if !input.trim().eq_ignore_ascii_case("y") {
-    //     println!("Update cancelled by user.");
-    //     return Ok(());
-    // }
-    //
-    // println!("\nUpdating Tracer to version {}...", latest_ver);
-
-    let install_cmd = if is_sudo_installed() {
-        "curl -sSL https://install.tracer.cloud/ | sudo bash && source ~/.bashrc && source ~/.zshrc"
-    } else {
-        if !is_root() {
-            println!("Warning: Running without root privileges. Some operations may fail.");
-        }
-        "curl -sSL https://install.tracer.cloud/ | bash && source ~/.bashrc && source ~/.zshrc"
-    };
-
-    let mut command = Command::new("bash");
-    command.arg("-c").arg(install_cmd);
-    let status = command
-        .status()
-        .context("Failed to update Tracer. Please try again.")?;
-
-    if !status.success() {
-        bail!("Failed to update Tracer. Please try again.");
-    }
-
-    println!(
-        "\n{} Tracer has been successfully updated!",
-        "Success:".green(),
-        // latest_ver
-    );
-
-    // println!(
-    //     "\n{} Tracer has been successfully updated to version {}!",
-    //     "Success:".green(),
-    //     latest_ver
-    // );
-    Ok(())
-}
